+++ functions/content.js
@@ -102,9 +102,9 @@ module.exports = async function content ({ page, context }) {
         return true;
       }, waitFor);
 
-      await (isSelector ? page.waitFor(waitFor) : page.evaluate(`(${waitFor})()`));
+      await (isSelector ? page.waitForSelector(waitFor) : page.evaluate(`(${waitFor})()`));
     } else {
-      await page.waitFor(waitFor);
+      await new Promise(r => setTimeout(r, waitFor));
     }
   }