+++ src/chrome-helper.ts
@@ -1,16 +1,21 @@
 import { ChildProcess } from 'child_process';
 // @ts-ignore no types
-import * as chromeDriver from 'chromedriver';
-import * as _ from 'lodash';
-import * as path from 'path';
-import * as puppeteer from 'puppeteer';
+import chromeDriver from 'chromedriver';
+import getPort from 'get-port';
+import _ from 'lodash';
+import path from 'path';
+import puppeteer from 'puppeteer';
+import pptrExtra from 'puppeteer-extra'
+import StealthPlugin from 'puppeteer-extra-plugin-stealth'
 import { ParsedUrlQuery } from 'querystring';
 import { Transform } from 'stream';
-import * as url from 'url';
+import treekill from 'tree-kill';
+import url from 'url';
+import { chromium, BrowserServer } from 'playwright-core';
 
 import { Features } from './features';
 import { browserHook, pageHook } from './hooks';
-import { fetchJson, getDebug, getUserDataDir, rimraf } from './utils';
+import { fetchJson, getDebug, getUserDataDir, rimraf, sleep } from './utils';
 
 import {
   IBrowser,
@@ -25,7 +30,9 @@ import {
 
 import {
   DEFAULT_BLOCK_ADS,
+  DEFAULT_DUMPIO,
   DEFAULT_HEADLESS,
+  DEFAULT_STEALTH,
   DEFAULT_IGNORE_DEFAULT_ARGS,
   DEFAULT_IGNORE_HTTPS_ERRORS,
   DEFAULT_LAUNCH_ARGS,
@@ -40,9 +47,9 @@ import {
   WORKSPACE_DIR,
 } from './config';
 
+import { PLAYWRIGHT_ROUTE } from './constants';
+
 const debug = getDebug('chrome-helper');
-const getPort = require('get-port');
-const treekill = require('tree-kill');
 const {
   CHROME_BINARY_LOCATION,
   USE_CHROME_STABLE,
@@ -68,6 +75,8 @@ const networkBlock = (request: puppeteer.Request) => {
 
 let runningBrowsers: IBrowser[] = [];
 
+pptrExtra.use(StealthPlugin());
+
 const parseIgnoreDefaultArgs = (query: ParsedUrlQuery): string[] | boolean => {
   const defaultArgs = query.ignoreDefaultArgs;
 
@@ -86,6 +95,10 @@ const parseIgnoreDefaultArgs = (query: ParsedUrlQuery): string[] | boolean => {
 
 const getTargets = async ({ port }: { port: string }): Promise<IJSONList[]> =>
   fetchJson(`http://127.0.0.1:${port}/json/list`);
+
+const isPuppeteer = (browserServer: puppeteer.Browser | BrowserServer): browserServer is puppeteer.Browser => {
+  return (browserServer as puppeteer.Browser).disconnect !== undefined;
+}
 
 const setupPage = async ({
   page,
@@ -143,6 +156,7 @@ const setupPage = async ({
 
 const setupBrowser = async ({
   browser: pptrBrowser,
+  browserWSEndpoint,
   isUsingTempDataDir,
   prebooted,
   browserlessDataDir,
@@ -152,9 +166,10 @@ const setupBrowser = async ({
   keepalive,
   process,
   windowSize,
-  port,
+  browserServer,
 }: {
   browser: puppeteer.Browser;
+  browserWSEndpoint: string;
   isUsingTempDataDir: boolean;
   browserlessDataDir: string | null;
   blockAds: boolean;
@@ -164,7 +179,7 @@ const setupBrowser = async ({
   keepalive: number | null;
   windowSize?: IWindowSize;
   prebooted: boolean;
-  port: number | string;
+  browserServer: BrowserServer | puppeteer.Browser;
 }): Promise<IBrowser> => {
   debug(`Chrome PID: ${process.pid}`);
   const browser = pptrBrowser as IBrowser;
@@ -180,15 +195,10 @@ const setupBrowser = async ({
   browser._prebooted = prebooted;
   browser._blockAds = blockAds;
   browser._pauseOnConnect = pauseOnConnect;
+  browser._browserServer = browserServer;
 
-  const { webSocketDebuggerUrl } = await fetchJson(`http://localhost:${port}/json/version`)
-    .catch((err) => {
-      closeBrowser(browser);
-      throw err;
-    });
-
-  browser._parsed = url.parse(webSocketDebuggerUrl, true);
-  browser._wsEndpoint = webSocketDebuggerUrl;
+  browser._parsed = url.parse(browserWSEndpoint, true);
+  browser._wsEndpoint = browserWSEndpoint;
   browser._id = (browser._parsed.pathname as string).split('/').pop() as string;
 
   await browserHook({ browser });
@@ -234,6 +244,8 @@ export const defaultLaunchArgs = {
   pauseOnConnect: false,
   slowMo: undefined,
   userDataDir: DEFAULT_USER_DATA_DIR,
+  playwright: false,
+  stealth: false,
 };
 
 /*
@@ -337,15 +349,27 @@ export const convertUrlParamsToLaunchOpts = (req: IHTTPRequest): ILaunchOptions 
     headless,
     ignoreHTTPSErrors,
     slowMo,
+    stealth,
     userDataDir,
     pause,
     trackingId,
     keepalive: keepaliveQuery,
+    dumpio: dumpioQuery,
   } = urlParts.query;
+
+  const playwright = req.parsed.pathname === PLAYWRIGHT_ROUTE;
 
   const isHeadless = !_.isUndefined(headless) ?
     headless !== 'false' :
     DEFAULT_HEADLESS;
+
+  const isStealth = !_.isUndefined(stealth) ?
+    stealth !== 'false' :
+    DEFAULT_STEALTH;
+
+  const dumpio = !_.isUndefined(dumpioQuery) ?
+    dumpioQuery !== 'false' :
+    DEFAULT_DUMPIO;
 
   const parsedKeepalive = _.parseInt(keepaliveQuery as string);
   const keepalive = _.isNaN(parsedKeepalive) ? undefined : parsedKeepalive;
@@ -354,11 +378,14 @@ export const convertUrlParamsToLaunchOpts = (req: IHTTPRequest): ILaunchOptions 
   return {
     args: !_.isEmpty(args) ? args : DEFAULT_LAUNCH_ARGS,
     blockAds: !_.isUndefined(blockAds) || DEFAULT_BLOCK_ADS,
+    dumpio,
     headless: isHeadless,
+    stealth: isStealth,
     ignoreDefaultArgs: parsedIgnoreDefaultArgs,
     ignoreHTTPSErrors: !_.isUndefined(ignoreHTTPSErrors) || DEFAULT_IGNORE_HTTPS_ERRORS,
     keepalive,
     pauseOnConnect: !_.isUndefined(pause),
+    playwright,
     slowMo: parseInt(slowMo as string, 10) || undefined,
     trackingId: _.isArray(trackingId) ? trackingId[0] : trackingId,
     userDataDir: userDataDir as string || DEFAULT_USER_DATA_DIR,
@@ -380,7 +407,10 @@ export const launchChrome = async (opts: ILaunchOptions, isPreboot: boolean): Pr
     executablePath: CHROME_BINARY_LOCATION,
     handleSIGINT: false,
     handleSIGTERM: false,
+    handleSIGHUP: false,
   };
+
+  const isPlaywright = launchArgs.playwright;
 
   // Having a user-data-dir in args is higher precedence than in opts
   const hasUserDataDir = _.some((launchArgs.args), (arg) => arg.includes('--user-data-dir='));
@@ -394,31 +424,63 @@ export const launchChrome = async (opts: ILaunchOptions, isPreboot: boolean): Pr
   }
 
   // If no data-dir is specified, use the default one in opts or generate one
+  // except for playwright which will error doing so.
   if (!hasUserDataDir) {
     browserlessDataDir = opts.userDataDir || await getUserDataDir();
     launchArgs.args.push(`--user-data-dir=${browserlessDataDir}`);
   }
 
-  // Only use debugging pipe when headless
+  // Only use debugging pipe when headless except for playwright which
+  // will error in doing so.
   if (isHeadless) {
     launchArgs.args.push(`--remote-debugging-pipe`);
   }
 
+  // Reset playwright to a workable state since it can't run headfull or use
+  // a user-data-dir
+  if (isPlaywright) {
+    launchArgs.args = launchArgs.args.filter((arg) => (
+        !arg.startsWith('--user-data-dir') &&
+        arg !== '--remote-debugging-pipe'
+      )
+    );
+    launchArgs.headless = true;
+  }
+
   debug(`Launching Chrome with args: ${JSON.stringify(launchArgs, null, '  ')}`);
 
-  return puppeteer.launch(launchArgs)
-    .then((browser: IBrowser) => setupBrowser({
+  const browserServer = launchArgs.playwright ?
+    await chromium.launchServer({
+      ...launchArgs,
+      headless: true,
+    }) :
+      launchArgs.stealth ?
+      await pptrExtra.launch(launchArgs):
+      await puppeteer.launch(launchArgs);
+
+  const { webSocketDebuggerUrl: browserWSEndpoint } = await fetchJson(`http://127.0.0.1:${port}/json/version`)
+    .catch((e) => {
+      browserServer.close();
+      throw e;
+    });
+
+  const iBrowser = isPuppeteer(browserServer) ?
+    Promise.resolve(browserServer) :
+    puppeteer.connect({ browserWSEndpoint })
+
+  return iBrowser.then((browser) => setupBrowser({
       blockAds: opts.blockAds,
       browser,
       browserlessDataDir,
+      browserWSEndpoint,
       isUsingTempDataDir,
       keepalive: opts.keepalive || null,
       pauseOnConnect: opts.pauseOnConnect,
-      process: browser.process(),
+      process: browserServer.process(),
       trackingId: opts.trackingId || null,
       windowSize: undefined,
       prebooted: isPreboot,
-      port,
+      browserServer,
     }));
 };
 
@@ -444,16 +506,16 @@ export const launchChromeDriver = async ({
 
         if (match) {
           chromeProcess.stderr && chromeProcess.stderr.unpipe(findPort);
-          const [, wsEndpoint] = match;
-          debug(`Attaching to chromedriver browser on ${wsEndpoint}`);
+          const [, browserWSEndpoint] = match;
+          debug(`Attaching to chromedriver browser on ${browserWSEndpoint}`);
 
-          const browser: puppeteer.Browser = await puppeteer.connect({ browserWSEndpoint: wsEndpoint });
-          const { port } = url.parse(wsEndpoint);
+          const browser: puppeteer.Browser = await puppeteer.connect({ browserWSEndpoint });
 
           iBrowser = await setupBrowser({
             blockAds,
             browser,
             browserlessDataDir,
+            browserWSEndpoint,
             isUsingTempDataDir,
             prebooted: false,
             keepalive: null,
@@ -461,7 +523,7 @@ export const launchChromeDriver = async ({
             process: chromeProcess,
             trackingId,
             windowSize,
-            port: port as string,
+            browserServer: browser,
           });
         }
 
@@ -517,13 +579,24 @@ export const closeBrowser = async (browser: IBrowser) => {
       rimraf(browser._browserlessDataDir);
     }
 
+    isPuppeteer(browser._browserServer) ? browser._browserServer.disconnect() : browser._browserServer.close();
     runningBrowsers = runningBrowsers.filter((b) => b._wsEndpoint !== browser._wsEndpoint);
-    browser.disconnect();
-    browser.removeAllListeners();
+
+    /*
+     * IMPORTANT
+     * You can't use close due to the possibility of a unhandled error event in
+     * a stream somewhere in puppeteer. Disconnect works, but doesn't cleanup events,
+     * so we're left with #disconnect + a manual removeAllListeners call and setting
+     * the browser object to `null` below to force it to collect
+     */
+    process.removeAllListeners('exit');
   } catch (error) {
     debug(`Browser close emitted an error ${error.message}`);
   } finally {
-    debug(`Sending SIGKILL signal to browser process ${browser._browserProcess.pid}`);
+    await sleep(1000);
+    debug(`Sending SIGKILL signal to browser process ${browser._browserProcess.pid}`);	
     treekill(browser._browserProcess.pid, 'SIGKILL');
+    // @ts-ignore force any garbage collection by nulling the browser
+    browser = null;
   }
 };