+++ src/hooks.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
-import * as fs from 'fs';
-import * as path from 'path';
+import fs from 'fs';
+import path from 'path';
 
 import {
   IBrowserHook,