+++ src/tests/integration/utils.ts
@@ -1,7 +1,7 @@
 import { exec as execNode } from 'child_process';
-import * as _ from 'lodash';
-import * as os from 'os';
-import * as util from 'util';
+import _ from 'lodash';
+import os from 'os';
+import util from 'util';
 import { IBrowserlessOptions } from '../../types';
 
 export const exec = util.promisify(execNode);
@@ -13,6 +13,7 @@ export const defaultParams = (): IBrowserlessOptions => ({
   disabledFeatures: [],
   enableAPIGet: true,
   enableCors: false,
+  enableHeapdump: false,
   errorAlertURL: null,
   exitOnHealthFailure: false,
   functionBuiltIns: ['url'],
@@ -30,10 +31,12 @@ export const defaultParams = (): IBrowserlessOptions => ({
   prebootChrome: false,
   queuedAlertURL: null,
   rejectAlertURL: null,
+  sessionCheckFailURL: null,
   singleRun: false,
   timeoutAlertURL: null,
   token: null,
   workspaceDir: os.tmpdir(),
+  socketBehavior: 'http',
 });
 
 export const throws = () => {