+++ src/tests/integration/websockets.spec.ts
@@ -1,4 +1,5 @@
-import * as puppeteer from 'puppeteer';
+import puppeteer from 'puppeteer';
+import { chromium } from 'playwright-core';
 
 import { BrowserlessServer } from '../../browserless';
 import { IBrowserlessOptions } from '../../types';
@@ -77,6 +78,45 @@ describe('Browserless Chrome WebSockets', () => {
       expect(browserless.currentStat.successful).toEqual(1);
       expect(browserless.currentStat.rejected).toEqual(0);
       expect(browserless.currentStat.queued).toEqual(0);
+      done();
+    });
+
+    job();
+  });
+
+  it('runs with no leaks', async (done) => {
+    const params = defaultParams();
+    const browserless = await start({
+      ...params,
+      connectionTimeout: -1,
+    });
+    await browserless.startServer();
+
+    const job = async () => {
+      return new Promise(async (resolve) => {
+        const browser: any = await puppeteer.connect({
+          browserWSEndpoint: `ws://127.0.0.1:${params.port}`,
+        });
+
+        browser.once('disconnected', resolve);
+
+        browser.disconnect();
+      });
+    };
+
+    browserless.queue.on('end', () => {
+      expect(browserless.currentStat.timedout).toEqual(0);
+      expect(browserless.currentStat.successful).toEqual(1);
+      expect(browserless.currentStat.rejected).toEqual(0);
+      expect(browserless.currentStat.queued).toEqual(0);
+
+      // browserless binds to these two events
+      // for graceful closing but puppeteer shouldn't
+      expect(process.listeners('SIGINT').length).toEqual(1);
+      expect(process.listeners('SIGTERM').length).toEqual(1);
+
+      expect(process.listeners('exit').length).toEqual(0);
+      expect(process.listeners('SIGHUP').length).toEqual(0);
       done();
     });
 
@@ -187,6 +227,27 @@ describe('Browserless Chrome WebSockets', () => {
       });
   });
 
+  it('fails requests with socket destroy', async () => {
+    const params = defaultParams();
+    const browserless = start({
+      ...params,
+      socketBehavior: 'close',
+      maxConcurrentSessions: 0,
+      maxQueueLength: 0,
+    });
+
+    await browserless.startServer();
+
+    return puppeteer.connect({ browserWSEndpoint: `ws://127.0.0.1:${params.port}` })
+      .then(throws)
+      .catch((error) => {
+        expect(browserless.currentStat.successful).toEqual(0);
+        expect(browserless.currentStat.rejected).toEqual(1);
+        expect(browserless.currentStat.queued).toEqual(0);
+        expect(error.message).toContain(`socket hang up`);
+      });
+  });
+
   it('fails requests in demo mode', async () => {
     const params = defaultParams();
     const browserless = start({
@@ -216,6 +277,81 @@ describe('Browserless Chrome WebSockets', () => {
     await browserless.startServer();
 
     return puppeteer.connect({ browserWSEndpoint: `ws://127.0.0.1:${params.port}` })
+      .then(throws)
+      .catch((error) => {
+        expect(browserless.currentStat.successful).toEqual(0);
+        expect(browserless.currentStat.rejected).toEqual(0);
+        expect(browserless.currentStat.queued).toEqual(0);
+        expect(error.message).toContain(`403`);
+      });
+  });
+
+  it('runs playwright', async (done) => {
+    const params = defaultParams();
+    const browserless = await start(params);
+    await browserless.startServer();
+
+    const job = async () => {
+      return new Promise(async (resolve) => {
+        const browser: any = await chromium.connect({
+          wsEndpoint: `ws://127.0.0.1:${params.port}/playwright`,
+        });
+
+        browser.once('disconnected', resolve);
+
+        browser.close();
+      });
+    };
+
+    browserless.queue.on('end', () => {
+      expect(browserless.currentStat.timedout).toEqual(0);
+      expect(browserless.currentStat.successful).toEqual(1);
+      expect(browserless.currentStat.rejected).toEqual(0);
+      expect(browserless.currentStat.queued).toEqual(0);
+      done();
+    });
+
+    job();
+  });
+
+  it(`doesn't allow playwright to do headfull or user-data-dirs`, async (done) => {
+    const params = defaultParams();
+    const browserless = await start(params);
+    await browserless.startServer();
+
+    const job = async () => {
+      return new Promise(async (resolve) => {
+        const browser: any = await chromium.connect({
+          wsEndpoint: `ws://127.0.0.1:${params.port}/playwright?--user-data-dir=/tmp&headless=false`,
+        });
+
+        browser.once('disconnected', resolve);
+
+        browser.close();
+      });
+    };
+
+    browserless.queue.on('end', () => {
+      expect(browserless.currentStat.timedout).toEqual(0);
+      expect(browserless.currentStat.successful).toEqual(1);
+      expect(browserless.currentStat.rejected).toEqual(0);
+      expect(browserless.currentStat.queued).toEqual(0);
+      done();
+    });
+
+    job();
+  });
+
+  it('rejects playwright without tokens', async () => {
+    const params = defaultParams();
+    const browserless = start({
+      ...params,
+      token: 'abc',
+    });
+
+    await browserless.startServer();
+
+    return chromium.connect({ wsEndpoint: `ws://127.0.0.1:${params.port}/playwright` })
       .then(throws)
       .catch((error) => {
         expect(browserless.currentStat.successful).toEqual(0);