+++ src/tests/util.spec.ts
@@ -1,6 +1,7 @@
 import { PassThrough } from 'stream';
-import * as utils from '../utils';
 import { IncomingMessage } from 'http';
+
+import * as utils from '../utils';
 
 const getArgs = (overrides = {}) => ({
   args: [],
@@ -11,6 +12,8 @@ const getArgs = (overrides = {}) => ({
   pauseOnConnect: false,
   slowMo: undefined,
   userDataDir: undefined,
+  playwright: false,
+  stealth: false,
   ...overrides,
 });